+++ cvd_src/fast/fast_12_detect.cxx
@@ -2253,7 +2253,7 @@ void fast_corner_detect_plain_12(const BasicImage<byte>& i, vector<ImageRef>& co
 			  continue;
 
 			success:
-				corners.push_back(ImageRef(cache_0-line_min, y));
+				corners.push_back(ImageRef(static_cast<int>(cache_0-line_min), y));
 		}
 	}
 }
